Created Jumping prototype scene
Created a jumping prototype scene and added a player charge jump mechanic. The longer the jump button/screen touch the longer the player will jump.

Also added a screenpoint to ray that takes into account where the player touches the screen on the x axis which affects the distance/direction of travel.